pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlsherlock"
version = "0.1.0"
description = "Intelligent ML agent CLI that diagnoses and fixes ML problems like a senior engineer"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1",
    "rich>=13.0",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "imbalanced-learn>=0.11",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
mlsh = "mlsherlock.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mlsherlock*"]
